Rename childFormField files to childFormFields for consistency, clean up
