+++ src/domain/registration/form/registrationForm.module.scss
@@ -1,6 +1,6 @@
 @import 'styles/variables';
 @import 'styles/layout';
-@import './partial/childFormField.module.scss';
+@import './partial/childFormFields.module.scss';
 
 $plusIconHeight: 2rem;
